rename terminal file
