+++ src/vs/workbench/contrib/terminal/browser/terminalView.ts
